+++ R/feed_author_feed.R
@@ -24,45 +24,54 @@
 bs_get_author_feed <- function(actor, cursor = NULL, limit = NULL,
                                user = get_bluesky_user(), pass = get_bluesky_pass(),
                                auth = bs_auth(user, pass), clean = TRUE) {
- if (missing(actor)) {
-   cli::cli_abort('{.arg actor} must list at least one user.')
- }
- if (!is.character(actor)) {
-   cli::cli_abort('{.arg actor} must be a character vector.')
- }
+  if (missing(actor)) {
+    cli::cli_abort('{.arg actor} must list at least one user.')
+  }
+  if (!is.character(actor)) {
+    cli::cli_abort('{.arg actor} must be a character vector.')
+  }
 
- if (!is.null(limit)) {
-   if (!is.numeric(limit)) {
-     cli::cli_abort('{.arg limit} must be numeric.')
-   }
-   limit <- as.integer(limit)
-   limit <- max(limit, 1L)
-   req_seq <- diff(unique(c(seq(0, limit, 100), limit)))
- } else {
-   req_seq <- list(NULL)
- }
+  if (!is.null(limit)) {
+    if (!is.numeric(limit)) {
+      cli::cli_abort('{.arg limit} must be numeric.')
+    }
+    limit <- as.integer(limit)
+    limit <- max(limit, 1L)
+    req_seq <- diff(unique(c(seq(0, limit, 100), limit)))
+  } else {
+    req_seq <- list(NULL)
+  }
 
- req <- httr2::request('https://bsky.social/xrpc/app.bsky.feed.getAuthorFeed') |>
-   httr2::req_url_query(actor = actor) |>
-   httr2::req_auth_bearer_token(token = auth$accessJwt) |>
-   httr2::req_url_query(
-     limit = limit
-   )
+  req <- httr2::request('https://bsky.social/xrpc/app.bsky.feed.getAuthorFeed') |>
+    httr2::req_url_query(actor = actor) |>
+    httr2::req_auth_bearer_token(token = auth$accessJwt) |>
+    httr2::req_url_query(
+      limit = limit
+    )
 
- resp <- repeat_request(req, req_seq, cursor, txt = 'Fetching followers')
+  resp <- repeat_request(req, req_seq, cursor, txt = 'Fetching followers')
 
- if (!clean) return(resp)
+  if (!clean) {
+    return(resp)
+  }
 
- resp |>
-   lapply(process_author_feed) |>
-   purrr::list_rbind() |>
-   add_req_url(req) |>
-   add_cursor(resp)
-}
-
-process_author_feed <- function(resp) {
-  resp |>
-    purrr::pluck('feed') |>
-    list_hoist() |>
+  lapply(resp,function(f) {
+    lapply(f$feed, function(x) {
+      if (is.null(x$reply)) {
+        proc_post(x$post)
+      } else{
+        dplyr::bind_cols(
+          proc_post(x$post),
+          widen(x$reply)
+        )
+      }
+    }) |>
+      purrr::list_rbind() |>
+      clean_names()
+  }) |>
+    purrr::list_rbind() |>
+    clean_names() |>
+    add_req_url(req) |>
+    add_cursor(resp) |>
     clean_names()
 }