+++ R/feed_get_posts.R
@@ -12,13 +12,14 @@
 #' @export
 #'
 #' @section Lexicon references:
-#' [feed/getPostThread.json (2023-10-01)](https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/feed/getPostThread.json)
+#' [feed/getPosts.json (2023-10-01)](https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/feed/getPosts.json)
 #'
 #' @section Function introduced:
 #' `v0.0.1` (2023-10-01)
 #'
 #' @examplesIf has_bluesky_pass() && has_bluesky_user()
 #' bs_get_posts('at://did:plc:ic6zqvuw5ulmfpjiwnhsr2ns/app.bsky.feed.post/3k7qmjev5lr2s')
+#' bs_get_posts('https://bsky.app/profile/chriskenny.bsky.social/post/3lc5d6zspys2c')
 bs_get_posts <- function(uris,
                          user = get_bluesky_user(), pass = get_bluesky_pass(),
                          auth = bs_auth(user, pass), clean = TRUE) {
@@ -29,7 +30,17 @@ bs_get_posts <- function(uris,
     cli::cli_abort('{.arg uris} must be a character vector.')
   }
 
-  uris <- uris |> as.list() |> purrr::set_names('uris')
+  uris <- purrr::map_chr(uris, function(x) {
+    if (stringr::str_detect(x, '^http')) {
+      bs_url_to_uri(x, auth = auth)
+    } else{
+      x
+    }
+  })
+
+  uris <- uris |>
+    as.list() |>
+    purrr::set_names('uris')
 
   req <- httr2::request('https://bsky.social/xrpc/app.bsky.feed.getPosts')
   req <- rlang::inject(httr2::req_url_query(req, !!!uris))
@@ -40,35 +51,15 @@ bs_get_posts <- function(uris,
     httr2::req_perform() |>
     httr2::resp_body_json()
 
-  if (!clean) return(resp)
+  if (!clean) {
+    return(resp)
+  }
 
-  resp <- resp |>
-    purrr::pluck('posts')
-
-  out <- lapply(resp, function(x) {
-    x |>
-      unlist(recursive = FALSE) |>
-      lapply(function(z) {
-        if (length(z) != 1) {
-          list(z)
-        } else {
-          z
-        }
-      }) |>
-      tibble::as_tibble_row()
-  }) |>
-    dplyr::bind_rows() |>
-    clean_names()
-
-  out <- out |>
-    dplyr::mutate(
-      dplyr::across(where(is.list) & where(~purrr::pluck_depth(.x) > 2),
-                    function(x) lapply(x, function(x) clean_names(proc(x))))
-    )
-
-  # out$record_embed <- lapply(out$record_embed, proc)
-  # out$embed_images <- lapply(out$embed_images, proc)
+  out <- resp |>
+    purrr::pluck('posts') |>
+    proc_posts()
 
   out |>
-    add_req_url(req)
+    add_req_url(req) |>
+    clean_names()
 }