+++ R/graph_actor_lists.R
@@ -21,10 +21,11 @@
 #'
 #' @examplesIf has_bluesky_pass() && has_bluesky_user()
 #' bs_get_actor_lists('profmusgrave.bsky.social')
+#' bs_get_actor_lists('chriskenny.bsky.social')
+#' bs_get_actor_lists('pfrazee.com')
 bs_get_actor_lists <- function(actor, cursor = NULL, limit = NULL,
                                user = get_bluesky_user(), pass = get_bluesky_pass(),
                                auth = bs_auth(user, pass), clean = TRUE) {
-
   if (missing(actor)) {
     cli::cli_abort('{.arg actor} must list at least one user.')
   }
@@ -50,7 +51,9 @@ bs_get_actor_lists <- function(actor, cursor = NULL, limit = NULL,
     )
   resp <- repeat_request(req, req_seq, cursor, txt = 'Fetching lists')
 
-  if (!clean) return(resp)
+  if (!clean) {
+    return(resp)
+  }
 
   resp |>
     lapply(process_actor_lists) |>