+++ R/record_post.R
@@ -3,18 +3,30 @@
 #' @param text text of post
 #' @param images character vector of paths to images to attach to post
 #' @param images_alt character vector of alt text for images. Must be same length as `images` if used.
+#' @param video character vector of path for up to one video to attach to post
+#' @param video_alt character vector, length one, of alt text for video, if used.
 #' @param langs character vector of languages in BCP-47 format
 #' @param reply character vector with link to the parent post to reply to
 #' @param quote character vector with link to a post to quote
+#' @param emoji boolean. Default is `TRUE`. Should `:emoji:` style references be converted?
+#' @param max_tries `r template_var_max_tries()`
 #' @param user `r template_var_user()`
 #' @param pass `r template_var_pass()`
 #' @param auth `r template_var_auth()`
 #' @param clean `r template_var_clean()`
 #'
+#' @details
+#' `:emoji:` parsing is not a formally supported Bluesky feature. This package
+#' converts usages of this kind by identifying text within `:`s, here "`emoji`"
+#' and then matches them to the `emoji` package's list of emoji names. All
+#' supported emoji names and corresponding images can be seen with
+#' `emoji::emoji_name`. This feature was introduced in `v0.2.0`.
+#'
+#'
 #' @concept record
 #'
 #' @section Lexicon references:
-#' [feed/post.json (2023-10-02)](https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/feed/post.json)
+#' [feed/post.json (2024-11-29)](https://github.com/bluesky-social/atproto/blob/main/lexicons/app/bsky/feed/post.json)
 #' [repo/createRecord.json (2023-10-02)](https://github.com/bluesky-social/atproto/blob/main/lexicons/com/atproto/repo/createRecord.json)
 #'
 #' @section Function introduced:
@@ -28,15 +40,28 @@
 #' via @bskyr.bsky.social (https://christophertkenny.com/bskyr/)')
 #' bs_post('Test self-reply from r package `bskyr`
 #' via @bskyr.bsky.social (https://christophertkenny.com/bskyr/)',
-#'         reply = 'https://bsky.app/profile/bskyr.bsky.social/post/3kexwuoyqj32g')
+#'   reply = 'https://bsky.app/profile/bskyr.bsky.social/post/3kexwuoyqj32g'
+#' )
 #' bs_post('Test quoting from r package `bskyr`
 #' via @bskyr.bsky.social (https://christophertkenny.com/bskyr/)',
-#'         quote = 'https://bsky.app/profile/bskyr.bsky.social/post/3kf24wd6cmb2a')
+#'   quote = 'https://bsky.app/profile/bskyr.bsky.social/post/3kf24wd6cmb2a'
+#' )
 #' bs_post('Test quote and reply from r package `bskyr`
 #' via @bskyr.bsky.social (https://christophertkenny.com/bskyr/)',
-#'         reply = 'https://bsky.app/profile/bskyr.bsky.social/post/3kexwuoyqj32g',
-#'         quote = 'https://bsky.app/profile/bskyr.bsky.social/post/3kf24wd6cmb2a')
-bs_post <- function(text, images, images_alt, langs, reply, quote,
+#'   reply = 'https://bsky.app/profile/bskyr.bsky.social/post/3kexwuoyqj32g',
+#'   quote = 'https://bsky.app/profile/bskyr.bsky.social/post/3kf24wd6cmb2a'
+#' )
+#'
+#' bs_post('Test quote with :emoji: and :fire: and :confetti_ball: from r package
+#'   `bskyr` via @bskyr.bsky.social (https://christophertkenny.com/bskyr/)')
+#'
+#' bs_post(text = 'testing sending videos from R',
+#'   video = fs::path_package('bskyr', 'man/figures/pkgs.mp4'),
+#'   video_alt = 'a carousel of package logos, all hexagonal')
+bs_post <- function(text, images, images_alt,
+                    video, video_alt,
+                    langs, reply, quote,
+                    emoji = TRUE, max_tries,
                     user = get_bluesky_user(), pass = get_bluesky_pass(),
                     auth = bs_auth(user, pass), clean = TRUE) {
   if (missing(text)) {
@@ -47,27 +72,64 @@ bs_post <- function(text, images, images_alt, langs, reply, quote,
     if (length(images) > 4) {
       cli::cli_abort('You can only attach up to 4 images to a post.')
     }
+
+    if (missing(images_alt) && !is.list(images)) {
+      cli::cli_abort('If {.arg images} is provided, {.arg images_alt} must also be provided.')
+    }
+  }
+
+  if (!missing(video)) {
+    if (length(video) > 1) {
+      cli::cli_abort('You can only attach one video to a post.')
+    }
+
+    if (missing(video_alt)) {
+      cli::cli_abort('If {.arg video} is provided, {.arg video_alt} must also be provided.')
+    }
+  }
+
+  if (!missing(images) && !missing(video)) {
+    cli::cli_abort('You can only attach images or a video to a post, not both.')
+  }
+
+  if (emoji) {
+    text <- parse_emoji(text)
   }
 
   facets_l <- parse_facets(txt = text, auth = auth)
 
   if (!missing(images)) {
-    if (is.list(images)) { #any(fs::path_ext(images) == '')
+    if (is.data.frame(images)) {
+      blob <- blob_tb_to_list(images)
+    } else if (is.list(images)) { # any(fs::path_ext(images) == '')
       # then we assume it's a blob
       blob <- images
     } else {
       # otherwise it's a set of paths
       blob <- bs_upload_blob(images, auth = auth, clean = FALSE)
-      if (length(blob) == 1) {
-        blob <- blob[[1]]
+    }
+
+    if (!missing(images_alt)) {
+      if (length(blob) != length(images_alt)) {
+        cli::cli_abort('{.arg images_alt} must be the same length as {.arg images}.')
       }
     }
   }
 
-  if (!missing(images_alt)) {
-    if (length(blob) != length(images_alt)) {
-      cli::cli_abort('{.arg images_alt} must be the same length as {.arg images}.')
+  if (!missing(video)) {
+    if (is.list(video)) {
+      # then we assume it's a blob
+      blob <- video
+    } else {
+      # otherwise it's a path
+      blob <- bs_upload_blob(video, auth = auth, clean = FALSE)
     }
+  }
+
+  if (stringi::stri_numbytes(text) > 300) {
+    # warning because *sometimes* it works even if I think it's above the limit
+    cli::cli_warn(c('{.arg text} evaluates to {stringi::stri_numbytes(text)} graphemes, which is above the limit (300).',
+                    i = 'If positng fails, consider reducing the length of the text.'))
   }
 
   post <- list(
@@ -86,9 +148,9 @@ bs_post <- function(text, images, images_alt, langs, reply, quote,
 
   if (!missing(images)) {
     if (!missing(images_alt)) {
-      img_incl <- lapply(seq_along(images), function(i) {
+      img_incl <- lapply(seq_along(blob), function(i) {
         list(
-          image  = blob[[i]]$blob,
+          image = blob[[i]]$blob,
           alt = images_alt[[i]]
         )
       })
@@ -104,6 +166,17 @@ bs_post <- function(text, images, images_alt, langs, reply, quote,
       '$type' = 'app.bsky.embed.images',
       images = img_incl
     )
+  }
+
+  if (!missing(video)) {
+    post$embed <- list(
+      '$type' = 'app.bsky.embed.video',
+      video = blob[[1]]$blob
+    )
+
+    if (!missing(video_alt)) {
+      post$embed$alt <- video_alt
+    }
   }
 
   if (!missing(reply)) {
@@ -137,16 +210,26 @@ bs_post <- function(text, images, images_alt, langs, reply, quote,
       )
     )
 
+  if (!missing(max_tries) && max_tries > 1) {
+    req <- req |>
+      httr2::req_retry(
+        max_tries = max_tries,
+        is_transient = function(x) httr2::resp_status(x) >= 400
+      )
+  }
+
   #return(httr2::req_dry_run(req))
 
   resp <- req |>
     httr2::req_perform() |>
     httr2::resp_body_json()
 
-  if (!clean) return(resp)
+  if (!clean) {
+    return(resp)
+  }
 
   resp |>
-    dplyr::bind_rows() |>
+    widen() |>
     clean_names() |>
     add_req_url(req)
 }