+++ R/repo_delete_record.R
@@ -30,12 +30,13 @@
 #' )
 #'
 #' rec <- bs_create_record(collection = 'app.bsky.feed.like', record = like)
-#' bs_delete_record(collection = 'app.bsky.feed.like',
-#' rkey = stringr::str_split_i(rec$uri, '/', i = 5))
+#' bs_delete_record(
+#'   collection = 'app.bsky.feed.like',
+#'   rkey = bs_extract_record_key(rec$uri)
+#' )
 bs_delete_record <- function(collection, rkey,
                              user = get_bluesky_user(), pass = get_bluesky_pass(),
                              auth = bs_auth(user, pass)) {
-
   req <- httr2::request('https://bsky.social/xrpc/com.atproto.repo.deleteRecord') |>
     httr2::req_auth_bearer_token(token = auth$accessJwt) |>
     httr2::req_body_json(
@@ -48,4 +49,7 @@ bs_delete_record <- function(collection, rkey,
 
   resp <- req |>
     httr2::req_perform()
+
+  resp |>
+    httr2::resp_status()
 }